Adiciona endpoint para obter vetor e calcula pontuação de fraude; ajusta configuração do Nginx e corrige versão no preprocess

+++ internal/preprocess/preprocess.go
@@ -16,7 +16,7 @@ const (
 	Magic      = "RVEC"
 	Dims       = 14
 	Stride     = 16
-	Version    = uint8(1)
+	Version    = uint32(1)
 	HeaderSize = 16
 )
